Refactor Kaldi compatibility tests
[ghstack-poisoned]
