+++ test/torchaudio_unittest/transforms/kaldi_compatibility_impl.py
@@ -1,6 +1,4 @@
 """Test suites for checking numerical compatibility against Kaldi"""
-import torch
-import torchaudio.functional as F
 import torchaudio.compliance.kaldi
 from parameterized import parameterized
 
@@ -22,22 +20,6 @@ class Kaldi(TempDirMixin, TestBaseMixin):
     def assert_equal(self, output, *, expected, rtol=None, atol=None):
         expected = expected.to(dtype=self.dtype, device=self.device)
         self.assertEqual(output, expected, rtol=rtol, atol=atol)
-
-    @skipIfNoExec('apply-cmvn-sliding')
-    def test_sliding_window_cmn(self):
-        """sliding_window_cmn should be numerically compatible with apply-cmvn-sliding"""
-        kwargs = {
-            'cmn_window': 600,
-            'min_cmn_window': 100,
-            'center': False,
-            'norm_vars': False,
-        }
-
-        tensor = torch.randn(40, 10, dtype=self.dtype, device=self.device)
-        result = F.sliding_window_cmn(tensor, **kwargs)
-        command = ['apply-cmvn-sliding'] + convert_args(**kwargs) + ['ark:-', 'ark:-']
-        kaldi_result = run_kaldi(command, 'ark', tensor)
-        self.assert_equal(result, expected=kaldi_result)
 
     @parameterized.expand(load_params('kaldi_test_fbank_args.json'))
     @skipIfNoExec('compute-fbank-feats')
